+++ packages/zero/vite.config.ts
@@ -55,7 +55,8 @@ export default defineLibConfig({
         'components/navbar/index': 'src/components/navbar/index.ts',
         'components/breadcrumbs/index': 'src/components/breadcrumbs/index.ts',
         'components/pagination/index': 'src/components/pagination/index.ts',
-        'components/steps/index': 'src/components/steps/index.ts'
+        'components/steps/index': 'src/components/steps/index.ts',
+        'components/drawer/index': 'src/components/drawer/index.ts'
     },
     external: ['sigx', 'sigx/jsx-runtime', 'sigx/jsx-dev-runtime'],
     jsx: true